cloud-run/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
google-generativeai==0.8.4
python-multipart==0.0.20
pydantic==2.10.4
